so fatigue

+++ ft_for_octal.c
@@ -6,37 +6,37 @@
 /*   By: rgwayne- <rgwayne-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/12 16:56:48 by rgwayne-          #+#    #+#             */
-/*   Updated: 2019/12/21 14:14:02 by rgwayne-         ###   ########.fr       */
+/*   Updated: 2020/01/14 21:09:44 by rgwayne-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 #include <limits.h>
 
-void ft_octal(t_struct *inform, int flag)
+void	ft_octal(t_struct *inform, int flag)
 {
-    int i;
+	int i;
 
-    i = 0;
-    if (flag == 3)
-        inform->value_d = (unsigned short)inform->value_d;
-    else if (flag == 4)
-        inform->value_d = (unsigned char)inform->value_d;
-    else if (flag == 5)
-        inform->value_d = (unsigned int)inform->value_d;
-    while (inform->value_d >= 8)
-    {
-        inform->value_hex[i] = inform->value_d % 8 + '0';
-        inform->value_d /= 8;
-        i++;
-    }
-    inform->value_hex[i] = inform->value_d + '0';
-    inform->final_size += i + 1;
+	i = 0;
+	if (flag == 3)
+		inform->value_d = (unsigned short)inform->value_d;
+	else if (flag == 4)
+		inform->value_d = (unsigned char)inform->value_d;
+	else if (flag == 5)
+		inform->value_d = (unsigned int)inform->value_d;
+	while (inform->value_d >= 8)
+	{
+		inform->value_hex[i] = inform->value_d % 8 + '0';
+		inform->value_d /= 8;
+		i++;
+	}
+	inform->value_hex[i] = inform->value_d + '0';
+	inform->final_size += i + 1;
 }
 
-int ft_value_o(t_struct *inform, va_list list, int i)
+int		ft_value_o(t_struct *inform, va_list list, int i)
 {
-    int len;
+	 int len;
 
     len = 0;
     if (i == 1)
@@ -47,7 +47,7 @@ int ft_value_o(t_struct *inform, va_list list, int i)
     }
 	else if (i == 2)
     {
-        inform->value_d = va_arg(list, unsigned long);
+        inform->value_d = va_arg(list, long long);
         ft_octal(inform, 2);
 		len = hex_len(inform->value_d, inform);
     }
